server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
ffmpeg-python==0.2.0
Pillow==11.1.0
aiofiles==24.1.0
